Refactor: Move IHasher and implement to Collections\ByteArrayConverter
