+++ src/Enums/Enum.php
@@ -7,8 +7,8 @@ use PHP\Collections\ByteArray;
 use PHP\Collections\Dictionary;
 use PHP\Enums\Exceptions\MalformedEnumException;
 use PHP\Hashing\HashAlgorithm\SHA256;
-use PHP\Hashing\Hasher\PrimitiveHasher;
-use PHP\Hashing\Hasher\SerializeThenApplyHashAlgorithm;
+use PHP\Collections\ByteArrayConverter\PrimitiveHasher;
+use PHP\Collections\ByteArrayConverter\SerializeThenApplyHashAlgorithm;
 use PHP\ObjectClass;
 use PHP\Serialization\PHPSerializer;
 use ReflectionClass;
