+++ tests/Collections/ByteArrayConverter/EquatableHasherTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Tests\Hashing\Hasher;
+namespace PHP\Tests\Collections\ByteArrayConverter;
 
 use PHP\Collections\ByteArray;
-use PHP\Hashing\Hasher\HasherDecorator;
-use PHP\Hashing\Hasher\IHasher;
-use PHP\Hashing\Hasher\EquatableHasher;
+use PHP\Collections\ByteArrayConverter\HasherDecorator;
+use PHP\Collections\ByteArrayConverter\IHasher;
+use PHP\Collections\ByteArrayConverter\EquatableHasher;
 use PHP\Interfaces\IEquatable;
 
 /**
