+++ tests/Collections/ByteArrayConverter/HasherDecoratorTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace PHP\Tests\Hashing\Hasher;
+namespace PHP\Tests\Collections\ByteArrayConverter;
 
 use PHP\Collections\ByteArray;
 use PHP\Exceptions\NotImplementedException;
-use PHP\Hashing\Hasher\HasherDecorator;
-use PHP\Hashing\Hasher\IHasher;
+use PHP\Collections\ByteArrayConverter\HasherDecorator;
+use PHP\Collections\ByteArrayConverter\IHasher;
 
 /**
  * Tests HasherDecorator
